# perf: improve initial load