--- a/desk/desk.bill
+++ b/desk/desk.bill
@@ -4,4 +4,5 @@
     %diary
     %hark
     %notify
+    %groups-ui
 ==