--- a/desk/mar/ui/migration.hoon
+++ b/desk/mar/ui/migration.hoon
@@ -1,0 +1,29 @@
+/-  u=ui
+::  group flag + channel flag
+|_  mig=migration:u
+++  grad  %noun
+++  grow
+  |%
+  ++  noun  mig
+  ++  json
+    =,  enjs:format
+    |^
+    %-  pairs
+    :~  chat-imports/(imported chat-imports.mig)
+        heap-imports/(imported heap-imports.mig)
+        diary-imports/(imported diary-imports.mig)
+        wait/a/(turn wait.mig |=(=ship s/(scot %p ship)))
+    ==
+    ++  imported
+      |=  =imported:u
+      %-  pairs
+      %+  turn  ~(tap by imported)
+      |=  [f=flag:g imp=?]
+      [(rap 3 (scot %p p.f) '/' q.f ~) b/imp]
+    --
+  --
+++  grab
+  |%
+  ++  noun  migration:u
+  --
+--