--- a/desk/sur/ui.hoon
+++ b/desk/sur/ui.hoon
@@ -1,0 +1,34 @@
+/-  g=groups, c=chat, d=diary, h=heap
+|%
++$  init
+  $%  =groups:g
+      =gangs:g
+      =chat
+      =heap
+      =diary    
+  ==
+::
++$  chat
+  $%  =briefs:c
+      chats=(map flag:c chat:c)
+  ==
+::
++$  heap
+  $%  =briefs:h
+      =stash:h
+  ==
+::
++$  diary
+  $%  =briefs:d
+      =shelf:d
+  ==
+::
++$  imported  (map flag:g ?)
+::
++$  migration
+  $%  chat-imports=imported
+      heap-imports=imported
+      diary-imports=imported
+      wait=(list ship)
+  ==
+--