--- a/talk/lib/chat-json.hoon
+++ b/talk/lib/chat-json.hoon
@@ -1,0 +1,1 @@
+../../desk/lib/chat-json.hoon