--- a/talk/lib/cite-json.hoon
+++ b/talk/lib/cite-json.hoon
@@ -1,0 +1,1 @@
+../../desk/lib/cite-json.hoon