--- a/talk/lib/groups-json.hoon
+++ b/talk/lib/groups-json.hoon
@@ -1,0 +1,1 @@
+../../desk/lib/groups-json.hoon