--- a/ui/src/logic/useMigrationInfo.ts
+++ b/ui/src/logic/useMigrationInfo.ts
@@ -3,21 +3,51 @@ import create from 'zustand';
 import api from '@/api';
 import { useGroupState } from '@/state/groups';
 import { GroupState } from '@/state/groups/type';
-import { useCallback, useEffect } from 'react';
+import { useCallback, useEffect, useMemo } from 'react';
 import useSchedulerStore from '@/state/scheduler';
+import { useChatState } from '@/state/chat';
+import { useHeapState } from '@/state/heap/heap';
+import { useDiaryState } from '@/state/diary';
 import { getNestShip, isChannelImported } from './utils';
-import usePendingImports from './usePendingImports';
+
+interface ChannelAgent {
+  pendingImports: Record<string, boolean>;
+}
+
+const selPendImports = (s: ChannelAgent) => s.pendingImports;
+
+export function usePendingImports() {
+  const chats = useChatState(selPendImports);
+  const heaps = useHeapState(selPendImports);
+  const diaries = useDiaryState(selPendImports);
+
+  return useMemo(
+    () => ({
+      ..._.mapKeys(chats, (v, c) => `chat/${c}`),
+      ..._.mapKeys(heaps, (v, h) => `heap/${h}`),
+      ..._.mapKeys(diaries, (v, d) => `diary/${d}`),
+    }),
+    [chats, heaps, diaries]
+  );
+}
+
+export interface MigrationInit {
+  wait: string[];
+  chat: Record<string, boolean>;
+  heap: Record<string, boolean>;
+  diary: Record<string, boolean>;
+}
 
 interface WaitStore {
   wait: string[];
-  fetchWait: () => Promise<void>;
+  fetchData: () => Promise<void>;
 }
 
 let initialized = false;
 const useWaitStore = create<WaitStore>((set, get) => ({
   initialized: false,
   wait: [],
-  fetchWait: async () => {
+  fetchData: async () => {
     if (initialized) {
       return;
     }
@@ -25,19 +55,15 @@ const useWaitStore = create<WaitStore>((set, get) => ({
     initialized = true;
 
     useSchedulerStore.getState().wait(async () => {
-      const wait = await api.scry<string[]>({
-        app: 'group-store',
-        path: '/wait',
+      const { wait, chat, heap, diary } = await api.scry<MigrationInit>({
+        app: 'groups-ui',
+        path: '/migration',
       });
 
+      useChatState.getState().initImports(chat);
+      useHeapState.getState().initImports(heap);
+      useDiaryState.getState().initImports(diary);
       set({ wait });
-      api.subscribe({
-        app: 'group-store',
-        path: '/wait',
-        event: (newWait: string[]) => {
-          set({ wait: newWait });
-        },
-      });
     }, 5);
   },
 }));
@@ -53,7 +79,7 @@ export function useStartedMigration(flag: string) {
   const pendingShips = Object.keys(pendingImports).map(getNestShip);
 
   useEffect(() => {
-    useWaitStore.getState().fetchWait();
+    useWaitStore.getState().fetchData();
   }, []);
 
   return {