--- a/ui/src/state/chat/type.ts
+++ b/ui/src/state/chat/type.ts
@@ -13,6 +13,7 @@ import {
   Clubs,
   Chats,
   ChatInit,
+  TalkChatInit,
 } from '../../types/chat';
 import { BaseState } from '../base';
 import { GroupMeta } from '../../types/groups';
@@ -22,9 +23,7 @@ export interface ChatState {
   batchSet: (fn: (sta: BasedChatState) => void) => void;
   chats: Chats;
   multiDms: Clubs;
-  dms: {
-    [ship: string]: Chat;
-  };
+  dms: string[];
   drafts: {
     [whom: string]: ChatStory;
   };
@@ -61,6 +60,7 @@ export interface ChatState {
   fetchPins: () => Promise<void>;
   markRead: (whom: string) => Promise<void>;
   start: (init: ChatInit) => Promise<void>;
+  startTalk: (init: TalkChatInit) => Promise<void>;
   dmRsvp: (ship: string, ok: boolean) => Promise<void>;
   getDraft: (whom: string) => void;
   fetchNewer: (ship: string, count: string) => Promise<boolean>;