--- a/ui/src/state/diary/type.ts
+++ b/ui/src/state/diary/type.ts
@@ -1,4 +1,3 @@
-import { Inline } from '@/types/content';
 import {
   Diary,
   DiaryBriefs,
@@ -31,6 +30,7 @@ export interface DiaryState {
   start: (init: DiaryInit) => Promise<void>;
   fetchNote: (flag: DiaryFlag, noteId: string) => Promise<void>;
   initialize: (flag: DiaryFlag) => Promise<void>;
+  initImports: (init: Record<string, boolean>) => void;
   joinDiary: (groupFlag: string, flag: DiaryFlag) => Promise<void>;
   leaveDiary: (flag: DiaryFlag) => Promise<void>;
   viewDiary: (flag: DiaryFlag, view: DiaryDisplayMode) => Promise<void>;