@jamesacklin 2023-03-10
~150% faster to load the app, and ~200% faster to load a chat. I'll take it!

As soon as we merge I'll throw this on canary so we can live in the new world for a bit.