@jamesacklin 2023-03-10
Fixes #1913 
Fixes #1375